Deleting task [2]
Previously we were differentiating with respect to the task name,but the issue was all the task name with same name would get deleted.

Hence,We need to have some identifier to differentiate each task in the list. Therefore we use -->id : for each element in the list

We consider each task as object consisting of taskName and id.

While adding,we cannot consider new value to be string(text),it should be an object.

So,we make the following changes :

In the addTask function,we create an object of id,and taskName,and create a list according to the ID.

Make the necessary changes in the main return statement as we are using the object consisting of id and taskName.

In the deleteTask function,we pass the ID,and compare the ID to delete the task.